initialize Schema Store
