+++ packages/schema-blocks/src/functions/redux/SchemaBlocksState.ts
@@ -1,5 +1,8 @@
 import { BlockValidationResult } from "../../core/validation";
 
+
+export const YOAST_SCHEMA_STORE_NAME = "yoast-seo/schema-blocks";
+
 export type SchemaBlocksState = {
 	validations: Record<string, BlockValidationResult>;
 };
