+++ packages/schema-blocks/src/functions/redux/index.ts
@@ -1,1 +1,4 @@
-export * from "./schemaBlocks";
+export * from "./actions";
+export * from "./reducer";
+export * from "./selectors";
+export * from "./initializeSchemaStore";
