+++ packages/schema-blocks/src/functions/redux/reducer/index.ts
@@ -1,1 +1,1 @@
-export * from "./schemaBlocks";
+export { schemaBlocksReducer } from "./schemaBlocks";
